Migrate API routes to client-side Firebase, remove test scripts
